+++ src/main/java/pieces/Position.java
@@ -1,4 +1,4 @@
-package chess;
+package pieces;
 
 public class Position {
