+++ src/test/java/chess/BoardTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import pieces.Piece;
+import pieces.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static pieces.Piece.Color;
