+++ src/test/java/pieces/PieceTest.java
@@ -1,6 +1,5 @@
 package pieces;
 
-import chess.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
